+++ strongtalk/src/cpp/main/vm/code/PolymorphicInlineCache.cpp
@@ -648,7 +648,7 @@ std::int32_t PolymorphicInlineCache::code_for_polymorphic_case( char *entry, Pol
         put_byte( p, MemOopDescriptor::klass_byte_offset() );
         st_assert( entry + static_cast<std::int32_t>( PolymorphicInlineCache::Constant::PolymorphicInlineCache_NativeMethod_entry_offset ) == p, "constant value inconsistent with code pattern" );
         // handle nativeMethods
-        for ( std::int32_t i = 0; i < c->n; i++ ) {
+        for ( std::size_t i = 0; i < c->n; i++ ) {
             // cmp edx, klass(i)
             st_assert( c->nativeMethod_klasses[ i ] not_eq smiKlassObject, "should not be smiKlassObject" );
             put_shrt( p, cmp_opcode );
@@ -874,7 +874,7 @@ void PolymorphicInlineCache::verify() {
     ResourceMark            rm;
     GrowableArray<KlassOop> *k = klasses();
 
-    for ( std::int32_t i = 0; i < k->length() - 1; i++ ) {
+    for ( std::size_t i = 0; i < k->length() - 1; i++ ) {
         for ( std::int32_t j = i + 1; j < k->length(); j++ ) {
             if ( k->at( i ) == k->at( j ) ) {
                 _console->print( "The class " );
